Extract answer_only_mode.rs from turn.rs (parent #680)
Create new sibling module `src/agent/loop_run/answer_only_mode.rs` to
host the answer-only-mode shell-command allowlist + script-execution
fallback response builder previously living in `turn.rs`. This is a
small follow-on extraction in the meta-issue #680 reduction effort —
the helpers form a clean responsibility-focused cluster but did not
fit one of the named Phase 1-8 modules.

Migrated to `answer_only_mode.rs` (`pub(super)`):

* `const ANSWER_ONLY_SCRIPT_ALLOWED_PREFIXES` — prefix whitelist
* `const ANSWER_ONLY_SCRIPT_BLOCKED_CONTAINS` — denylist (shell
    operators + file-mutating commands)
* `const ANSWER_ONLY_SCRIPT_BLOCKED_PREFIXES` — prefix denylist for
    file-mutating commands
* `fn truncate_for_answer(text, max_chars) -> String` (UTF-8 safe,
    keep `max_chars - 32` chars, append `\n...[truncated N chars]`
    footer)
* `fn answer_only_script_execution_fallback_response(output) -> String`
    (Japanese templated response showing trimmed output + exit-code
    summary)
* `fn answer_only_script_command_allowed(command) -> bool` (top-level
    predicate, walks `cd <dir> && <rest>` chains via
    module-private helpers)
* Module-private supporting helpers (`answer_only_script_contains_any`
    / `_starts_with_any` / `answer_only_cd_segment_allowed` /
    `answer_only_cd_chained_script_tail` /
    `answer_only_script_has_blocked_operation`)

`loop_run.rs` adjustments:

* Registered `mod answer_only_mode;` with a parent-#680 doc comment
  documenting the responsibility split.

`turn.rs` adjustments:

* Removed the const + 7 fn definitions (originally at lines 534-612).
* Added the 2 caller-facing imports
  (`answer_only_script_command_allowed` /
  `answer_only_script_execution_fallback_response`) via a new
  `use super::answer_only_mode::{...}` block.
* Test mod call sites (`super::answer_only_script_command_allowed` /
  `super::answer_only_script_execution_fallback_response`) resolve
  via `super::*` (re-export from `super` mod scope; tests previously
  consumed the same names through the `tests` mod's pre-existing
  `super::{...}` import block).

LOC delta:

* `turn.rs`: 28,809 → 28,732 (-77 LOC)
* `answer_only_mode.rs`: 0 → 107 (new)
* Cumulative `turn.rs`: 39,489 → 28,732 (-10,757 LOC, -27.2%)

`pub(super)` limited / no facade re-export (DR3-001).
`turn.rs` remains the only in-crate consumer.

Verification:

* `cargo fmt`
* `cargo clippy --lib --all-targets -- -D warnings` clean
* `cargo test --lib` 3,114 passed / 0 failed
